chore: Clean up and split shuffle module
